issue #1 - Syntax Highlighting
Syntax now gets highlighted according to the selected mode.
The pastes are not yet saved anywhere but simply routed to pasted.ejs
with a POST.

Since a few modes use the same highlighting file (clike.js)
 - Java
 - C/C++
 - C#
we would need to add a MIME-type for each of those modes.
http://codemirror.net/mode/clike/index.html

This is not yet implemented and reserved for later, since the main goal
of this ticket has been achieved.