Restored hotkey functionality
To restore the hotkeys I brought in mousetrap and moustrap
global. index.html now pulls in these scripts. In the process of
adding mousetrap I moved all of the script tags in index.html to just
before the close of body as this has become an industry best practice.

In refheap.js and create.js I exposed the codemirror editor and the
refheap module on the window object. This was necessary to keep all of
the hotkey binding in one place.

resolves macskay/pasteit#3